upd: (utils/) {setboxes,share/install_symlinks}.sh; install symlinks only if they do not exist.

+++ utils/setboxes.sh
@@ -1,4 +1,8 @@
 #!/bin/sh
+if [ ! -L "${DESTDIR}/mss/work" ]; then
+    ln -sf /mnt/mss/stuff/techy-bits/work "${DESTDIR}/mss/work"
+fi
 
-ln -sf /mnt/mss/stuff/techy-bits/work "${DESTDIR}/mss/work"
-ln -sf /mnt/mss/stuff/techy-bits/git/setboxes "${DESTDIR}/mss/repo" 
+if [ ! -L "${DESTDIR}/mss/repo" ]; then
+    ln -sf /mnt/mss/stuff/techy-bits/git/setboxes "${DESTDIR}/mss/repo"
+fi
